# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "gestioneventos",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/ubicaciones": {
            "get": {
                "tags": [
                    "Ubicaciones"
                ],
                "summary": "Mostrar todas las ubicaciones",
                "responses": {
                    "200": {
                        "description": "Lista de ubicaciones",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Ubicacion"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No se encontraron ubicaciones",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No se encontró ninguna ubicación con esa capacidad."
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Ubicaciones"
                ],
                "summary": "Agregar una nueva ubicación",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre_ubicacion": {
                                        "type": "string",
                                        "example": "Sala A"
                                    },
                                    "capacidad": {
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Ubicación principal para eventos"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ubicación creada con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Ubicacion"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error en la validación de los datos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error en la validación de los datos"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 400
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/ubicaciones/{id}": {
            "get": {
                "tags": [
                    "Ubicaciones"
                ],
                "summary": "Mostrar una ubicación específica",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ubicación encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Ubicacion"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ubicación no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ubicacion no encontrada"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Ubicaciones"
                ],
                "summary": "Actualizar una ubicación",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre_ubicacion": {
                                        "type": "string",
                                        "example": "Sala A"
                                    },
                                    "capacidad": {
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Ubicación principal para eventos"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ubicación actualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Ubicacion"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ubicación no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ubicacion no encontrada"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error en la validación de los datos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error en la validación de los datos"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 400
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Ubicaciones"
                ],
                "summary": "Eliminar una ubicación",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ubicación eliminada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ubicacion eliminada"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ubicación no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ubicacion no encontrada"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Ubicaciones"
                ],
                "summary": "Actualizar parcialmente una ubicación",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre_ubicacion": {
                                        "type": "string",
                                        "example": "Sala A"
                                    },
                                    "capacidad": {
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Ubicación principal para eventos"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ubicación actualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Ubicacion"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ubicación no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ubicacion no encontrada"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error en la validación de los datos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error en la validación de los datos"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 400
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Ubicacion": {
                "title": "Ubicacion",
                "description": "Modelo de la ubicación",
                "properties": {
                    "id_ubicacion": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre_ubicacion": {
                        "type": "string",
                        "example": "Sala A"
                    },
                    "capacidad": {
                        "type": "integer",
                        "example": 100
                    },
                    "descripcion": {
                        "type": "string",
                        "example": "Ubicación principal para eventos"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}